requirements.txt
asgiref==3.10.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==5.0
idna==3.11
jwcrypto==1.5.6
pycparser==2.23
PyJWT==2.8.0
requests==2.32.5
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
